Deprecate helm check
While kube-score and kubeval were fitting nicely our purposes in the past, this is no longer the case.
kubeval isn't supported anymore, meaning that we would have to find an alternative either way to support our current 1.21 version. Kube-score was misbeaving in certain scenarios, giving a false sense of safety to the engineers.

We need to take a step back and find better alternatives.